more work to split stalkerutils out
