+++ stalkeragent/requirements.txt
@@ -1,2 +1,3 @@
 eventlet==0.17.4
 greenlet==0.4.9
+stalkerutils==2.0.2
