+++ stalkerutils/.gitignore
@@ -1,0 +1,3 @@
+build
+dist
+*.egg-info
