+++ stalkerutils/README.md
@@ -1,0 +1,4 @@
+Stalker Monitoring System - PoC
+===============================
+
+# Collection of utilities for stalker-agent and stalker-web
